Use REDUC_GROUP_FIRST_ELEMENT less
REDUC_GROUP_FIRST_ELEMENT is often checked to see whether we are
dealing with a SLP reduction or a reduction chain.  When we are
in the context of analyzing the reduction (so we are sure
the SLP instance we see is correct), then we can use the SLP
instance kind instead.

	* tree-vect-loop.cc (get_initial_defs_for_reduction): Adjust
	comment.
	(vect_create_epilog_for_reduction): Get at the reduction
	kind via the instance, re-use the slp_reduc flag instead
	of checking REDUC_GROUP_FIRST_ELEMENT again.
	Remove unreachable code.
	(vectorizable_reduction): Compute a reduc_chain flag from
	the SLP instance kind, avoid REDUC_GROUP_FIRST_ELEMENT
	checks.
	(vect_transform_cycle_phi): Likewise.
	(vectorizable_live_operation): Check the SLP instance
	kind instead of REDUC_GROUP_FIRST_ELEMENT.