rtl-ssa: Add missing live-out uses [PR121619]
This PR is another bug in the rtl-ssa code to manage live-out uses.
It seems that this didn't get much coverage until recently.

In the testcase, late-combine first removed a register-to-register
move by substituting into all uses, some of which were in other EBBs.
This was done after checking make_uses_available, which (as expected)
says that single dominating definitions are available everywhere
that the definition dominates.  But the update failed to add
appropriate live-out uses, so a later parallelisation attempt
tried to move the new destination into a later block.

gcc/
	PR rtl-optimization/121619
	* rtl-ssa/functions.h (function_info::commit_make_use_available):
	Declare.
	* rtl-ssa/blocks.cc (function_info::commit_make_use_available):
	New function.
	* rtl-ssa/changes.cc (function_info::apply_changes_to_insn): Use it.

gcc/testsuite/
	PR rtl-optimization/121619
	* gcc.dg/pr121619.c: New test.