libstdc++: Specialize _Never_valueless_alt for jthread, stop_token and stop_source
The move constructors for stop_source and stop_token are equivalent to
copying and clearing the raw pointer, as they are wrappers for a
counted-shared state.

For jthread, the move constructor performs a member-wise move of stop_source
and thread. While std::thread could also have a _Never_valueless_alt
specialization due to its inexpensive move (only moving a handle), doing
so now would change the ABI. This patch takes the opportunity to correct
this behavior for jthread, before C++20 API is marked stable.

libstdc++-v3/ChangeLog:

	* include/std/stop_token (__variant::_Never_valueless_alt): Declare.
	(__variant::_Never_valueless_alt<std::stop_token>)
	(__variant::_Never_valueless_alt<std::stop_source>): Define.
	* include/std/thread: (__variant::_Never_valueless_alt): Declare.
	(__variant::_Never_valueless_alt<std::jthread>): Define.