Introduce abstraction for vect reduction info, tracked from SLP nodes
While we have already the accessor info_for_reduction, its result
is a plain stmt_vec_info.  The following turns that into a class
for the purpose of changing accesses to reduction info to a new
set of accessors prefixed with VECT_REDUC_INFO and removes
the corresponding STMT_VINFO prefixed accessors where possible.

There is few reduction related things that are used by scalar
cycle detection and thus have to stay as-is for now and as
copies in future.

This also separates reduction info into one object per reduction
and associate it with SLP nodes, splitting it out from
stmt_vec_info, retaining (and duplicating) parts used by scalar
cycle analysis.  The data is then associated with SLP nodes
forming reduction cycles and accessible via info_for_reduction.
The data is created at SLP discovery time as we look at it even
pre-vectorizable_reduction analysis, but most of the data is
only populated by the latter.  There is no reduction info with
nested cycles that are not part of an outer reduction.
In the process this adds cycle info to each SLP tree, notably
the reduc-idx and a way to identify the reduction info.

	* tree-vectorizer.h (vect_reduc_info): New.
	(create_info_for_reduction): Likewise.
	(VECT_REDUC_INFO_TYPE): Likewise.
	(VECT_REDUC_INFO_CODE): Likewise.
	(VECT_REDUC_INFO_FN): Likewise.
	(VECT_REDUC_INFO_SCALAR_RESULTS): Likewise.
	(VECT_REDUC_INFO_INITIAL_VALUES): Likewise.
	(VECT_REDUC_INFO_REUSED_ACCUMULATOR): Likewise.
	(VECT_REDUC_INFO_INDUC_COND_INITIAL_VAL): Likewise.
	(VECT_REDUC_INFO_EPILOGUE_ADJUSTMENT): Likewise.
	(VECT_REDUC_INFO_FORCE_SINGLE_CYCLE): Likewise.
	(VECT_REDUC_INFO_RESULT_POS): Likewise.
	(VECT_REDUC_INFO_VECTYPE): Likewise.
	(STMT_VINFO_VEC_INDUC_COND_INITIAL_VAL): Remove.
	(STMT_VINFO_REDUC_EPILOGUE_ADJUSTMENT): Likewise.
	(STMT_VINFO_FORCE_SINGLE_CYCLE): Likewise.
	(STMT_VINFO_REDUC_FN): Likewise.
	(STMT_VINFO_REDUC_VECTYPE): Likewise.
	(vect_reusable_accumulator::reduc_info): Adjust.
	(vect_reduc_type): Adjust.
	(_slp_tree::cycle_info): New member.
	(SLP_TREE_REDUC_IDX): Likewise.
	(vect_reduc_info_s): Move/copy data from ...
	(_stmt_vec_info): ... here.
	(_loop_vec_info::redcu_infos): New member.
	(info_for_reduction): Adjust to take SLP node.
	(vect_reduc_type): Adjust.
	(vect_is_reduction): Add overload for SLP node.
	* tree-vectorizer.cc (vec_info::new_stmt_vec_info):
	Do not initialize removed members.
	(vec_info::free_stmt_vec_info): Do not release them.
	* tree-vect-stmts.cc (vectorizable_condition): Adjust.
	* tree-vect-slp.cc (_slp_tree::_slp_tree): Initialize
	cycle info.
	(vect_build_slp_tree_2): Compute SLP reduc_idx and store
	it.  Create, populate and propagate reduction info.
	(vect_print_slp_tree): Print cycle info.
	(vect_analyze_slp_reduc_chain): Set cycle info on the
	manual added conversion node.
	(vect_optimize_slp_pass::start_choosing_layouts): Adjust.
	* tree-vect-loop.cc (_loop_vec_info::~_loop_vec_info):
	Release reduction infos.
	(info_for_reduction): Get the reduction info from
	the vector in the loop_vinfo.
	(vect_create_epilog_for_reduction): Adjust.
	(vectorizable_reduction): Likewise.
	(vect_transform_reduction): Likewise.
	(vect_transform_cycle_phi): Likewise, deal with nested
	cycles not part of a double reduction have no reduction info.
	* config/aarch64/aarch64.cc (aarch64_force_single_cycle):
	Use VECT_REDUC_INFO_FORCE_SINGLE_CYCLE, get SLP node and use
	that.
	(aarch64_vector_costs::count_ops): Adjust.