OpenMP: Improve front-end error-checking for "declare variant"
This patch fixes a number of problems with parser error checking of
"declare variant", especially in the C front end.

The new C testcase unprototyped-variant.c added by this patch used to
ICE when gimplifying the call site, at least in part because the
variant was being recorded even after it was diagnosed as invalid.
There was also a large block of dead code in the C front end that was
supposed to fix up an unprototyped declaration of a variant function
to match the base function declaration, that was never executed because
it was nested in a conditional that could never be true.  I've fixed those
problems by rearranging the code and only recording the variant if it
passes the correctness checks.  I also tried to add some comments and
re-work some particularly confusing bits of code, so that it's easier to
understand.

The OpenMP specification doesn't say what the behavior of "declare
variant" with the "append_args" clause should be when the base
function is unprototyped.  The additional arguments are supposed to be
inserted between the last fixed argument of the base function and any
varargs, but without a prototype, for any given call we have no idea
which arguments are fixed and which are varargs, and therefore no idea
where to insert the additional arguments.  This used to trigger some
other diagnostics (which one depending on whether the variant was also
unprototyped), but I thought it was better to just reject this with an
explicit "sorry".

Finally, I also observed that a missing "match" clause was only
rejected if "append_args" or "adjust_args" was present.  Per the spec,
"match" has the "required" property, so if it's missing it should be
diagnosed unconditionally.  The C++ and Fortran front ends had the same
issue so I fixed this one there too.

gcc/c/ChangeLog
	* c-parser.cc (c_finish_omp_declare_variant): Rework diagnostic
	code.  Do not record variant if there are errors.  Make check for
	a missing "match" clause unconditional.

gcc/cp/ChangeLog
	* parser.cc (cp_finish_omp_declare_variant): Structure diagnostic
	code similarly to C front end.  Make check for a missing "match"
	clause unconditional.

gcc/fortran/ChangeLog
	* openmp.cc (gfc_match_omp_declare_variant): Make check for a
	missing "match" clause unconditional.

gcc/testsuite/ChangeLog
	* c-c++-common/gomp/append-args-1.c: Adjust expected output.
	* g++.dg/gomp/adjust-args-1.C: Likewise.
	* g++.dg/gomp/adjust-args-3.C: Likewise.
	* gcc.dg/gomp/adjust-args-1.c: Likewise:
	* gcc.dg/gomp/append-args-1.c: Likewise.
	* gcc.dg/gomp/unprototyped-variant.c: New.
	* gfortran.dg/gomp/adjust-args-1.f90: Adjust expected output.
	* gfortran.dg/gomp/append_args-1.f90: Likewise.