tree-optimization/121638 - missed SLP discovery of live induction
The following fixes a missed SLP discovery of a live induction.
Our pattern matching of those fails because of the PR81529 fix
which I think was misguided and should now no longer be relevant.
So this essentially reverts that fix.  I have added a GIMPLE
testcase to increase the chance the particular IL is preserved
through the future.

This shows that how we make some IVs live because of early-break
isn't quite correct, so I had to preserve a hack here.  Hopefully
to be investigated at some point.

	PR tree-optimization/121638
	* tree-vect-stmts.cc (process_use): Do not make induction
	PHI backedge values relevant.

	* gcc.dg/vect/pr121638.c: New testcase.