arm: testsuite: make gcc.target/arm/bics_3.c generate bics again
The compiler is getting too smart!  But this test is really intended
to test that we generate BICS instead of BIC+CMP, so make the test use
something that we can't subsequently fold away into a bit minipulation
of a store-flag value.

I've also added a couple of extra tests, so we now cover both the
cases where we fold the result away and where that cannot be done.
Also add a test that we don't generate a compare against 0, since
that's really part of what this test is covering.

gcc/testsuite:

	* gcc.target/arm/bics_3.c: Add some additional tests that
	cannot be folded to a bit manipulation.