More RISC-V testsuite hygiene
More testsuite hygiene.  Some of the thead tests are expecting to find
xtheadvdot in the extension set, but it's not defined as a valid extension
anywhere.  I'm just removing xtheadvdot.  Someone more familiar with these
cores can add it back properly if they're so inclined.

Second, there's a space after the zifencei in a couple of the thead arch
strings.  Naturally that causes failures as well.  That's a trivial fix, just
remove the bogus whitespace.

That gets us clean on riscv.exp on the pioneer system.

The pioneer is happy, as is riscv32-elf and riscv64-elf.  Pushing to the trunk.

gcc/
	* config/riscv/riscv-cores.def (xt-c908v): Drop xtheadvdot.
	(xt-c910v2): Remove extraenous whitespace.
	(xt-c920v2): Drop xtheadvdot and remove extraeonous whitespace.

gcc/testsuite/

	* gcc.target/riscv/mcpu-xt-c908v.c: Drop xtheadvdot.
	* gcc.target/riscv/mcpu-xt-c920v2.c: Drop xtheadvdot.