tree-optimization/121686 - failed SLP discovery for live recurrence
The following adjusts the SLP build for only-live stmts to not
only consider vect_induction_def and vect_internal_def that are
not part of a reduction but instead consider all non-reduction
defs that are not part of a reduction, specifically in this case
a recurrence def.  This is also a missed optimization on the
gcc-15 branch (but IMO a very minor one).

	PR tree-optimization/121686
	* tree-vect-slp.cc (vect_analyze_slp): Consider all only-live
	non-reduction defs for discovery.

	* gcc.dg/vect/pr121686.c: New testcase.