ada: Disable ghost context checks before context is set
There are cases where we need to analyze the argument of the pragma
in order to determine the ghostliness of the pragma. However during
that analysis the ghost region of the pragma is not set yet so we
cannot perform the ghost context checks at that moment.

This patch provides the mechanism for disabling ghost context checks
and disables them for pragma arguments that determine the ghostliness
of the pragma.

gcc/ada/ChangeLog:

	* ghost.adb (Check_Ghost_Context): Avoid context checks when they
	are globally disabled.
	* sem.ads (Ghost_Context_Checks_Disabled): New flag to control
	whether ghost context checks are activated or not.
	* sem_prag.adb (Analyze_Pragma): Disable ghost context checks for
	pragmas that determine their ghostliness based on one of its arguments.