c++/modules: Add explanatory note for incomplete types with definition in different module [PR119844]
The confusion in the PR arose because the definition of 'User' in a
separate named module did not provide an implementation for the
forward-declaration in the global module.  This seems likely to be a
common mistake while people are transitioning to modules, so this patch
adds an explanatory note.

While I was looking at this I also noticed that the existing handling of
partial specialisations for this note was wrong (we pointed at the
primary template declaration rather than the relevant partial spec), so
this patch fixes that up, and also gives a more precise error message
for using a template other than by self-reference while it's being
defined.

	PR c++/119844

gcc/cp/ChangeLog:

	* typeck2.cc (cxx_incomplete_type_inform): Add explanation when
	a similar type is complete but attached to a different module.
	Also fix handling of partial specs and templates.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/pr119844_a.C: New test.
	* g++.dg/modules/pr119844_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>